Extract editor placement preview helpers
