+++ src/game_data_internal.h
@@ -1055,6 +1055,9 @@ void init_editor_selection(slayer3d_game_data_editor_selection *selection);
 bool editor_selection_active_for_scene(const slayer3d_game_data_runtime *runtime);
 bool editor_command_preview_active_for_scene(const slayer3d_game_data_runtime *runtime);
 bool editor_placement_preview_active_for_scene(const slayer3d_game_data_runtime *runtime);
+void clear_editor_placement_preview(slayer3d_game_data_runtime *runtime);
+void update_editor_placement_preview(slayer3d_game_data_runtime *runtime, yyjson_val *editor,
+                                     const slayer3d_game_data_editor_selection *hover_selection);
 void editor_set_string_output(slayer3d_properties *props, yyjson_val *outputs, const char *key_name, const char *value);
 void editor_set_bool_output(slayer3d_properties *props, yyjson_val *outputs, const char *key_name, bool value);
 void editor_set_int_output(slayer3d_properties *props, yyjson_val *outputs, const char *key_name, int value);
